Feat: 기존 프로필 이미지가 없는 경우에는 S3 삭제 요청하지 않도록 수정

+++ src/user/user.service.ts
@@ -127,7 +127,7 @@ export class UserService {
       throw new UserNotFoundException();
     }
 
-    if (oldImageKey !== newImageKey) {
+    if (oldImageKey && oldImageKey !== newImageKey) {
       await this.s3Service.deleteFile(oldImageKey);
     }
   }
